Refactored Alarm class into its own base-level class. Also implemented public methods to set Acknowledgements and Shelving states and associated metadata.
Fixed certain errors that appeared as a result of this refactoring.

Also removed existing ControlPanel and renamed this as ControlPanelBeta - this can be freely removed at a later point, for the moment it is there for archiving purposes. Re-created ControlPanel.java class fresh to better meet the requirements for the Panel. 

Also added a method in the TreeItemSelector abstract superclass to change the selection method for the JTree used as part of it.

Finally, implemented the SimpleTreeSelector class at a basic level so as to be able to add a simple Tree and retrieve selections without the large table displaying the selections.