Changed default size of MonFrame.
Updated layout of AlarmManagerPanel with some new default behaviour, particularly on the display panel.

Made some small fixes to ControlPanel the scrollbar behaviour when using a mousewheel.

Added a new method to the PointSourceSelector to enable the automatic selection of all points when none have been added to the table.