Updates to Alarm, AlarmPanel and AlarmManagerPanel - small bug fixes and layout changes.
Ensured that the automatic updates in AlarmManagerPanel don't switch the tab if the alarm in question is actually shelved or acknowledged already.

Changed the look of AlarmPanel so that if the username that acknowledged or shelved the alarm is null, the word "null" and "ASAP" is not displayed. Also set the displayed time to be in UTC rather than a hexadecimal reference that it is set to by default.

Added code so that the guidance text (if present in the PointDescription) is properly populated when creating an Alarm object