Refactored the ignoreList into the AlarmMaintainer so it is no centrally maintained and persists through a single MoniCA client instance, effectively mirroring the behaviour expected from other point states.
Also changed the behaviour of the "Dismiss" button from the popup alarm panels - this now adds it to a temporary ignore list as well as the "real" ignore list. When this point goes out of its active alarming state, the point is removed from the ignore list so simply dismissing it means that the user is not continuously bothered by an alarm popup, but it returns to regular behaviour once it drops out of its alarming state, either through acknowledgement, shelving or natural loss of alarming status.